requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
alembic==1.16.2
altair==5.5.0
altgraph==0.17.2
annotated-types==0.7.0
anthropic==0.69.0
anyio==3.7.1
asgiref==3.8.1
# Removed async-timeout explicit pin - version conflict with langchain 0.3.25
# async-timeout will be installed as dependency of aiohttp/asyncpg/langchain
# All require compatible versions: aiohttp needs >=4.0, asyncpg needs >=4.0.3, langchain needs <5.0.0
# pip will automatically resolve to 4.0.3+ which satisfies all constraints
asyncpg==0.29.0
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==23.12.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.0.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.8
coloredlogs==15.0.1
coverage==7.10.7
cryptography==45.0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
docstring-parser==0.17.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.5.1
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.2
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
macholib==1.15.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
networkx==3.2.1
nodeenv==1.9.1
notion-client==2.5.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.23.2
openai==1.82.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pathspec==0.12.1
Pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==4.2.0
pre_commit==4.2.0
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pulsar-client==3.7.0
pyarrow==20.0.0
Pillow==11.3.0
platformdirs==4.3.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.9.1
pydantic_core==2.18.4
pydeck==0.9.1
PyGithub==1.59.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.45.1
structlog==23.2.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
google-generativeai>=0.3.0
